+++ webapp/src/main/resources/static/js/search-channel-page/searchChannel.js
