+++ webapp/src/main/resources/templates/search-channel-page.html
@@ -5,8 +5,8 @@
     <meta charset="UTF-8">
     <title>Workspace Page</title>
     <script th:src="@{/js/jquery-3.4.1.min.js}"></script>
-    <script type="module" th:src="@{/js/searchChannel.js}"></script>
-    <link rel="stylesheet" th:href="@{/css/searchChannel.css}"/>
+    <script type="module" th:src="@{/js/search-channel-page/searchChannel.js}"></script>
+    <link rel="stylesheet" th:href="@{/js/search-channel-page/searchChannel.css}"/>
 </head>
 
 <body>
