+++ webapp/src/main/resources/templates/search-users-page.html
@@ -5,8 +5,8 @@
     <meta charset="UTF-8">
     <title>Workspace Page</title>
     <script th:src="@{/js/jquery-3.4.1.min.js}"></script>
-    <script type="module" th:src="@{/js/searchUsers.js}"></script>
-    <link rel="stylesheet" th:href="@{/css/searchChannel.css}"/>
+    <script type="module" th:src="@{/js/search-users-page/searchUsers.js}"></script>
+    <link rel="stylesheet" th:href="@{/js/search-users-page/searchUsers.css}"/>
 </head>
 
 <body>
