+++ webapp/src/main/resources/templates/workspace-page.html
@@ -6,7 +6,7 @@
     <title>Workspace Page</title>
     <link rel="stylesheet" th:href="@{../bootstrap/bootstrap-4.3.1-dist/css/bootstrap.min.css}"/>
     <script th:src="@{/js/jquery-3.4.1.min.js}"></script>
-    <script th:src="@{/js/workspace-page.js}" type="module"></script>
+    <script th:src="@{/js/workspace-page-page/workspace-page.js}" type="module"></script>
     <script th:src="@{../bootstrap/bootstrap-4.3.1-dist/js/bootstrap.min.js}"></script>
     <script defer th:src="@{/js/input-box.js}" type="module"></script>
     <script defer th:src="@{/js/input-box-at.js}" type="module"></script>
@@ -31,9 +31,9 @@
     <link rel="stylesheet" th:href="@{/css/button.css}"/>
     <script src="/js/displayDiv.js"></script> -->
 
-    <link rel="stylesheet" th:href="@{/css/workspace-page.css}"/>
-    <link rel="stylesheet" th:href="@{/css/workspace-page-old.css}"/>
-    <link rel="stylesheet" th:href="@{/css/workspace-add-channel-modal.css}"/>
+    <link rel="stylesheet" th:href="@{/js/workspace-page-page/workspace-page.css}"/>
+    <link rel="stylesheet" th:href="@{/js/workspace-page-page/workspace-page-old.css}"/>
+    <link rel="stylesheet" th:href="@{/js/workspace-page-page/workspace-add-channel-modal.css}"/>
     <script src="https://cdnjs.cloudflare.com/ajax/libs/sockjs-client/1.4.0/sockjs.min.js"
             integrity="sha256-z4cKT3yF+afIn8eFXJc+nRpGtwcmNEClvf+ikgsZcRo=" crossorigin="anonymous"></script>
     <script src="https://cdnjs.cloudflare.com/ajax/libs/stomp.js/2.3.3/stomp.min.js"
